requirements.txt
asgiref==3.2.7
boto3==1.14.43
botocore==1.17.43
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.3.0
django-storages==1.9.1
djangorestframework==3.11.0
dnspython==2.0.0
docutils==0.15.2
futures==3.1.1
gunicorn==20.0.4
jmespath==0.10.0
olefile==0.46
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.10

